+++ lib/neighbor/type/vector.rb
@@ -5,10 +5,9 @@ module Neighbor
         :vector
       end
 
-      # TODO uncomment in 0.4.0
-      # def deserialize(value)
-      #   value[1..-1].split(",").map(&:to_f) unless value.nil?
-      # end
+      def deserialize(value)
+        value[1..-1].split(",").map(&:to_f) unless value.nil?
+      end
     end
   end
 end