+++ test/neighbor_test.rb
@@ -225,22 +225,22 @@ class NeighborTest < Minitest::Test
   def test_type
     if vector?
       Item.create!(factors: "[1,2,3]")
-      assert_equal "[1,2,3]", Item.last.factors
+      assert_equal [1, 2, 3], Item.last.factors
 
       Item.create!(factors: [1, 2, 3])
-      assert_equal "[1,2,3]", Item.last.factors
+      assert_equal [1, 2, 3], Item.last.factors
     else
       Item.create!(factors: "(1,2,3)")
-      assert_equal "(1, 2, 3)", Item.last.factors
+      assert_equal [1, 2, 3], Item.last.factors
 
       Item.create!(factors: [1, 2, 3])
-      assert_equal "(1, 2, 3)", Item.last.factors
+      assert_equal [1, 2, 3], Item.last.factors
 
       Item.create!(factors: 1)
-      assert_equal "(1)", Item.last.factors
+      assert_equal [1], Item.last.factors
 
       Item.create!(factors: [[1, 2, 3], [4, 5, 6]])
-      assert_equal "(1, 2, 3),(4, 5, 6)", Item.last.factors
+      assert_equal [[1, 2, 3], [4, 5, 6]], Item.last.factors
     end
   end