+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for `halfvec` type
 - Added support for `taxicab`, `hamming`, and `jaccard` distances with `vector` extension
+- Added deserialization for `cube` and `vector` columns without `has_neighbor`
 - Fixed issue with `nearest_neighbors` scope overriding `select` values
 - Dropped support for Ruby < 3.1